+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/CommentLikeUserEntity.cs
@@ -5,16 +5,18 @@ namespace PnP.Core.Model.SharePoint
     /// <summary>
     /// userEntity class, write your custom code here
     /// </summary>
-    [SharePointType("Microsoft.SharePoint.Likes.userEntity", Uri = "_api/xxx", LinqGet = "_api/xxx")]
-    internal partial class CommentUserEntity : BaseDataModel<ICommentUserEntity>, ICommentUserEntity
+    [SharePointType("Microsoft.SharePoint.Likes.userEntity")]
+    internal partial class CommentLikeUserEntity : BaseDataModel<ICommentLikeUserEntity>, ICommentLikeUserEntity
     {
         #region Construction
-        public CommentUserEntity()
+        public CommentLikeUserEntity()
         {
         }
         #endregion
 
         #region Properties
+
+        public DateTime CreationDate { get => GetValue<DateTime>(); set => SetValue(value); }
 
         [SharePointProperty("email")]
         public string Mail { get => GetValue<string>(); set => SetValue(value); }
@@ -28,9 +30,6 @@ namespace PnP.Core.Model.SharePoint
         [KeyProperty(nameof(Id))]
         public override object Key { get => Id; set => Id = (int)value; }
 
-        #endregion
-
-        #region Extension methods
         #endregion
     }
 }