+++ src/sdk/PnP.Core/Model/SharePoint/Pages/Internal/Page.cs
@@ -47,7 +47,7 @@ namespace PnP.Core.Model.SharePoint
         #endregion
 
         #region Page Properties
-        
+
         /// <summary>
         /// Title of the client side page
         /// </summary>
@@ -253,7 +253,7 @@ namespace PnP.Core.Model.SharePoint
             List<IPage> loadedPages = new List<IPage>();
 
             // Get a reference to the pages library, reuse the existing one if the correct properties were loaded
-            IList pagesLibrary = await EnsurePagesLibraryAsync(context).ConfigureAwait(false);            
+            IList pagesLibrary = await EnsurePagesLibraryAsync(context).ConfigureAwait(false);
 
             // Prepare CAML query to load the list items
             await GetPagesListData(pageName, pagesLibrary).ConfigureAwait(false);
@@ -1715,7 +1715,7 @@ namespace PnP.Core.Model.SharePoint
             if ((layoutType == PageLayoutType.Article || LayoutType == PageLayoutType.Spaces) && PageListItem[PageConstants.BannerImageUrlField] != null)
             {
                 if (string.IsNullOrEmpty(PageListItem[PageConstants.BannerImageUrlField].ToString()) ||
-                    ((PageListItem[PageConstants.BannerImageUrlField] is FieldUrlValue bannerImageUrlFieldValue) && 
+                    ((PageListItem[PageConstants.BannerImageUrlField] is FieldUrlValue bannerImageUrlFieldValue) &&
                       bannerImageUrlFieldValue.Url.Contains("/_layouts/15/images/sitepagethumbnail.png", StringComparison.InvariantCultureIgnoreCase)))
                 {
                     string previewImageServerRelativeUrl = "";
@@ -1814,7 +1814,7 @@ namespace PnP.Core.Model.SharePoint
 
             try
             {
-                pageFile = await pagesLibrary.PnPContext.Web.GetFileByServerRelativeUrlAsync($"{pagesLibrary.RootFolder.ServerRelativeUrl}/{pageName}", p => p.ListItemAllFields, p => p.ServerRelativeUrl).ConfigureAwait(false);
+                pageFile = await pagesLibrary.PnPContext.Web.GetFileByServerRelativeUrlAsync($"{pagesLibrary.RootFolder.ServerRelativeUrl}/{pageName}", p => p.ListItemAllFields, p => p.ServerRelativeUrl, p => p.ListId).ConfigureAwait(false);
             }
             catch (SharePointRestServiceException ex)
             {
@@ -1841,6 +1841,15 @@ namespace PnP.Core.Model.SharePoint
                 (var folderName, var pageNameWithoutFolder) = PageToPageNameAndFolder(pageName);
 
                 PageListItem = pageFile.ListItemAllFields;
+
+                // plug in the pages library list id in metadata to help with url token resolving
+                if (PageListItem is IMetadataExtensible metadataExtensible)
+                {
+                    if (!metadataExtensible.Metadata.ContainsKey(PnPConstants.MetaDataListId))
+                    {
+                        metadataExtensible.Metadata.Add(PnPConstants.MetaDataListId, pageFile.ListId.ToString());
+                    }
+                }
 
                 if (!PageListItem.Values.ContainsKey(PageConstants.FileDirRef))
                 {
@@ -2024,7 +2033,7 @@ namespace PnP.Core.Model.SharePoint
         {
             if (PageListItem != null)
             {
-                return await PnPContext.Web.GetFileByServerRelativeUrlAsync($"{PageListItem[PageConstants.FileDirRef]}/{PageListItem[PageConstants.FileLeafRef]}", expressions).ConfigureAwait(false);                
+                return await PnPContext.Web.GetFileByServerRelativeUrlAsync($"{PageListItem[PageConstants.FileDirRef]}/{PageListItem[PageConstants.FileLeafRef]}", expressions).ConfigureAwait(false);
             }
             else
             {
@@ -2032,7 +2041,7 @@ namespace PnP.Core.Model.SharePoint
             }
         }
 
-        public void Publish(string comment=null)
+        public void Publish(string comment = null)
         {
             PublishAsync(comment).GetAwaiter().GetResult();
         }
@@ -2051,19 +2060,19 @@ namespace PnP.Core.Model.SharePoint
                     var sitePagesLibrary = await EnsurePagesLibraryAsync(PnPContext).ConfigureAwait(false);
 
                     if (pageFile.CheckOutType != CheckOutType.None)
-                    { 
+                    {
                         // Needs checkin
                         await pageFile.CheckinAsync(comment, sitePagesLibrary.EnableMinorVersions ? CheckinType.MinorCheckIn : CheckinType.MajorCheckIn).ConfigureAwait(false);
                     }
 
                     if (sitePagesLibrary.EnableMinorVersions)
-                    { 
+                    {
                         // Publishing
                         await pageFile.PublishAsync(comment).ConfigureAwait(false);
                     }
 
                     if (sitePagesLibrary.EnableModeration)
-                    { 
+                    {
                         // Approval 
                         await pageFile.ApproveAsync(comment).ConfigureAwait(false);
                     }
@@ -2150,7 +2159,7 @@ namespace PnP.Core.Model.SharePoint
         }
         #endregion
 
-        #region Page comment handling
+        #region Page commenting and liking
         public bool AreCommentsDisabled()
         {
             return AreCommentsDisabledAsync().GetAwaiter().GetResult();
@@ -2160,7 +2169,7 @@ namespace PnP.Core.Model.SharePoint
         {
             if (PageListItem != null)
             {
-                return await PageListItem.AreCommentsDisabledAsync().ConfigureAwait(false);                
+                return await PageListItem.AreCommentsDisabledAsync().ConfigureAwait(false);
             }
             else
             {
@@ -2211,6 +2220,78 @@ namespace PnP.Core.Model.SharePoint
 
             await PageListItem.SetCommentsDisabledAsync(true).ConfigureAwait(false);
         }
+
+        public ICommentCollection GetComments(params Expression<Func<IComment, object>>[] selectors)
+        {
+            return GetCommentsAsync(selectors).GetAwaiter().GetResult();
+        }
+
+        public async Task<ICommentCollection> GetCommentsAsync(params Expression<Func<IComment, object>>[] selectors)
+        {
+            // ensure we do have the page list item loaded
+            if (PageListItem == null)
+            {
+                await EnsurePageListItemAsync(pageName).ConfigureAwait(false);
+            }
+
+            return await PageListItem.GetCommentsAsync(selectors).ConfigureAwait(false);
+        }
+
+        public async Task LikeAsync()
+        {
+            await LikeUnlike(true).ConfigureAwait(false);
+        }
+
+        public void Like()
+        {
+            LikeAsync().GetAwaiter().GetResult();
+        }
+
+        public async Task UnlikeAsync()
+        {
+            await LikeUnlike(false).ConfigureAwait(false);
+        }
+
+        public void Unlike()
+        {
+            UnlikeAsync().GetAwaiter().GetResult();
+        }
+
+        private async Task LikeUnlike(bool like)
+        {
+            // ensure we do have the page list item loaded
+            if (PageListItem == null)
+            {
+                await EnsurePageListItemAsync(pageName).ConfigureAwait(false);
+            }
+
+            if (like)
+            {
+                await (PageListItem as ListItem).LikeAsync().ConfigureAwait(false);
+            }
+            else
+            {
+                await (PageListItem as ListItem).UnlikeAsync().ConfigureAwait(false);
+            }
+        }
+
+        public async Task<ILikedByInformation> GetLikedByInformationAsync()
+        {
+            // ensure we do have the page list item loaded
+            if (PageListItem == null)
+            {
+                await EnsurePageListItemAsync(pageName).ConfigureAwait(false);
+            }
+
+            // Already load the actual likes, assuming this will be needed in most cases and thus saving the roundtrip
+            return (await PageListItem.GetAsync(p => p.LikedByInformation.QueryProperties(p => p.LikeCount, p => p.IsLikedByUser, p => p.LikedBy)).ConfigureAwait(false)).LikedByInformation;
+        }
+
+        public ILikedByInformation GetLikedByInformation()
+        {
+            return GetLikedByInformationAsync().GetAwaiter().GetResult();
+        }
+
         #endregion
 
         #region Page Deletion